+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- No changes yet.
+### Fixed
+- yarpcerrors: `fmt` verbs are ignored when no args are passed to error
+  constructors.
 
 ## [1.45.0] - 2020-04-21
 ### Added